rflib: Introduce TLV, Action, Match, Option types
TLV is a type/length/value structure that supports converting to/from
BSON. It has constructors based on a pointer or different sized integer
values.

Match,Action,Option are subtypes of TLV that restrict the values to be
stored inside the TLV to specific types and values. They also provide
functions to handle converting vectors of those types to/from BSON.

Signed-off-by: Joe Stringer <joestringernz@gmail.com>