RouteMod: Fix route deletion
The pair of (vm_id,port) is required in RFServer to determine which datapath
route updates should be sent to. Previously we didn't send the port with
delete messages, so they were being dropped at RFServer.

This patch includes the port number when sending route removal
RouteMods, allowing routes to be removed from a datapath. RFServer will
use the output action to determine which datapath to forward the message
to, then drop the output action from the RouteMod.

Signed-off-by: Joe Stringer <joestringernz@gmail.com>