Update README.md
Moved the build.sh instructions to the end of the page and brought back the previous ones.
Some newcomers were having trouble with build.sh. The instructions as they are now  will be considered a first step, and the build.sh script will be targeted to developers who need more advanced features.